requirements.txt
numpy
Pillow
PyYAML
scikit-image
fastapi
uvicorn
python-multipart
